Add `syncgithub.sh` script
Synchronizes the local ~/GitHub directory with everything it thinks you
have access to on GitHub. Requires a few things to be put in your
gitconfig file(s) so it has an OAuth token to work with and a backup
system to point to. Useful for people like me who want access to the
projects they work on when GitHub is down. I run this once a week? Could
automate it, but whatevs.