Create new `close-cmd.sh` script for git
Figures out what branch you want to close (expecting the first argument
to be either a branch or GitHub issue number), and then merges it into
the current branch.

If given an issue number, auto-detects the branch if it is named e.g.
`issue/123` or `hotfix/123`.

Automatically generates a merge commit message indicating the issue
being closed; e.g., closing issue `123` causes the commit message body
to include `Closes #123` so that the issue gets closed when the merge
commit is pushed to GitHub. Also includes the `--gpg-sign` flag, so that
the merge commit is signed.