Update gpg-agent config and init to work on linux boxes?
Had been playing with changing the `pinentry-program` entry in
gpg-agent.conf and creating a second socket for ssh to connect through;
gave up and just changed how gpg-agent is started in allshrc.sh. Now
selects the socket location and decides how to start gpg-agent based on
whether or not `uname` outputs `Darwin`.