Add commit counts to logrepos script output
Shows how many repositories have been touched since the specified time,
and gives a listing of how many such commits there are for each
repository. As with the rest of the script, this expects that commits
made by the current user will have the `user.name` value from the global
.gitconfig file in each commit's `author` metadata. If the author name
differs sometimes, e.g. when committing locally vs committing on GitHub,
the results will not be accurate.