Rework git repository status display on login; reformat, add stash count
Now includes a header, and formats each line of output as:

Clean, [master], [  0] -- dotfiles
Dirty, [  #123], [ 12] -- dotfiles

Current columns are work-tree status (clean or dirty), current branch,
and stash count. As before, if none of these points are worth showing
for a specific repository, that repository is not shown; and issue
branches are detected if their branch name looks like 'issue/123'.