Rework the git status display again
Moved most of the code into a function to make it a bit easier to
troubleshoot issues, and changed how the stash count is detected to
speed it up a little bit. Tried changing out how the branch name is
detected, which worked okay, but did not provide any (noticable)
speedup.