Update show_git_status shell function to handle release branches
Any repo on a branch matching `release/.*` will now be listed, and the
release number will be prettified. If the branch matches `release/v.*`,
then everything after the `/` will be shown; otherwise, the `v` will be
prepended.

Also adjusted the padding on branch names to 8 characters, and adjusted
non-issue and non-release branch names to be left-aligned.