Update install.sh to behave slightly better on Win10 WSL
For when one wishes to run zsh in bash in ubuntu in windows...

At any rate, giving up on actually doing so for the moment, because I do
not seem to be able to get my gpg-agent setup working. My Win10
environment is a virtual machine on a macOS system, and I use a PGP
smartcard to authenticate against ssh, so my choices are to either
forward the USB device to the Win10 VM and lose ssh ability on the host,
or find a way to forward the agent somehow.