Add .mutt directory with muttrc
Points off to `~/.local/mutt/...` for the `mailcap` file because the
`html2text` utility will be in a different location on some of my
systems, and points to the same directory for an `account_config` file
so that each system can have its own accounts set up.

Fairly minimal config so far, just poking around and trying out
different configuration values.