Add mutt macro for converting messages from markdown to HTML
Was trying to write up a shell one-liner that I could put into the macro
command, but had some trouble and gave up. New `markdown.sh` script does
a bunch of piping around to make sure that the signature keeps the
correct formatting, and uses `cmark` to do the actual conversion.