Add new `todo.sh` script
Creates a few text files under `~/.local/share/todo_list/` and uses them
as a stack. Enter some text, and it gets added as a new entry on "top"
of the stack (actually at the bottom of the text file, and therefore
also the bottom of the display). Although not shown, each task includes
the time at which it was added to the stack.

If you "pop" the current task, or mark it as "done" (enter either of
those or one of 'p' or 'd'), the current task is removed from the task
list, and appended to the "done" file. The contents of the "done" file
can be viewed with either "finished" or "f". Each line in the finished
tasks list contains the task start time, the task finish time, and the
task contents.

Additionally, enter a task number, and that task will become the current
task. Alternatively, enter "p" followed by (or preceded-by) a task
number, and that task will be marked as finished, and moved to the
appropriate file.

Currently have no way to completely remove a task from either list, but
they are both stored as text files, so just open them with a text editor
if this is needed for now?