Add `dashboard` function to allshrc.sh
Opens tmux and does a bunch of splits to arrange my monitor a
pre-defined set of panes. One shows the current calendar and the weather
for the current location (set per-system in ~/.local/.allshrc), and
another shows the current output of `show_git_status`; these two panes
refresh every 5 minutes. Another pane is opened to the `todo.sh` script
in this repository. The two remaining panes are intended to be used for
mail via mutt, and for short commands into the terminal.