Adjust all committable task changes to specify task name/number
This means that in the long-term, the git commit messages duplicate the
information contained in the text files, but means that the commit
messages are easier to look at in `git log` output. Keeping the original
text files in place at the same time prevents me from having to parse
through the commit messages, which is good because someone (me) might
just go making random commits in their todo-list repository.