Add `--color` to the git `lg1` alias
git 2.14.2 changed pretty-print formatting color codes to not show color
if not in a tty; this means that when I run logrepos.sh, the output is
no longer colored.

Adding `--color` to the `git log` call enforces that no, we really do
want ANSI escape codes for colors.

The relevant git release notes section:

 * "%C(color name)" in the pretty print format always produced ANSI
   color escape codes, which was an early design mistake.  They now
   honor the configuration (e.g. "color.ui = never") and also tty-ness
   of the output medium.