Add un-released commit count to show_git_status function
Shown as "UnRel" in the display, this shows the number of non-merge
commits which have occurred in the current branch which are not also
present in `origin/master`. For me, this will highlight repositories
which have had work done and merged to dev recently since the last
release, which at some point need to be wrapped up as a new release.