Adjust syncgithub.sh to add an `upstream` remote to forked repos
Before trying to sync with each remote, checks to see if `origin` exists
and is a GitHub repository; if it is, uses the same
`github.access-token` setting used for initial GitHub repository
discovery, and gets GitHub's data on this repository. If their API
claims that this repository is a fork, gets the SSH URL, removes the
"git@" from the front (so that it matches the SSH URL format that I
use), and adds it as a new `upstream` remote. Sync then continues
normally, ensuring that the new upstream remote is synchronized.