Set merge.conflictstyle to diff3 in my global git config
Adding this because @eevee said to

Also because I read the documentation, and it sounds like a good thing
to have turned on. It adds a `|||||||` line to merge conflicts,
immediately followed by the original value, so that we have just a
little bit more context as to what's going on during a merge conflict.