Add RUNNING_GIT_CLOSE env var to the merge in `git close`
Running on WSL, I need to use `gpg.exe` for authentication against SSH
servers, and for signing commits; however, using the Windows copy of GPG
does not work with git's signature status info on commits. Added a local
bash script in my WSL home directory that calls `gpg.exe` when this
environment variable is set (so that the `--gpg-sign` option on the
merge can use my smart card), and uses the local `gpg` otherwise. Seems
to work so far!