Turn off `allowrevins` in neovim, because I don't actually write hebrew?
Turned on this option in my vimrc ages ago while perusing the docs for
settings to toggle, and must not have realized that this not only turns
on reverse indent, but also sets vim to insert in hebrew? This can be
toggled with `hkmap`, but realistically I'm not going to type in reverse
very often, and probably don't need a keyboard shortcut for it.
Meanwhile, on <company laptop>, I keep hitting Ctrl and Shift at the
same time when I go to type underscores, and suddenly I'm typing weird
characters with umlauts and such. Better to turn this off and then
manually set `revins` if I want to type in reverse.