Adjust syncgithub script to work on my work WSL environment
In my newer work machine's WSL environment, the clone URI needs to be a
bit different to use SSH via `plink.exe` (so that one of my Yubikeys can
be used for SSH authentication), so the script needed a few changes to
handle this. At some point I also experimented with using port 443 and
`ssh.github.com` to clone from GitHub, so there's some changes here for
handling that as well.

Although I doubt I will remember that this is here, a note for myself
and anyone else who feels like reading it: when using PuTTY and it's
"plink" tool for lots of short-lived connections, you may want to set it
up to maintain sessions. With ssh, this can be done using the
`ControlMaster` and `ControlPath` settings, and then running e.g.
`ssh git@github.com -N`. The `-N` parameter is required when connecting
to GitHub this way, as it tells ssh not to open a tty for you to type
into - GitHub doesn't allow tty access, but you want the connection to
stay open. Once you've run a command like this once for each server you
are going to connect to, run whatever batch script you're trying to run.

When using PuTTY and plink, to do the same, you need to open PuTTY and
change the `Default Settings` session. Load that session, go into
Connection > SSH, and check the `Share SSH connections if possible`
option, then save the `Default Settings` session. Once that is done,
instead of running ssh commands for each server, run plink like so:
`plink.exe git@github.com -N`. The `-N` parameter does the same thing
here as it does with ssh.