Add initial configuration for Windows Terminal
Still trying to figure out how to just symlink this to the right
location, but having some trouble making it happen.

Copied colors from the theme I was using in ConEmu, and added a WSL
profile as my default since that's where I do most of my command-line
work. Included a fair amount of padding to make up for one of their
known UI bugs where text is getting chopped off at the edges of the
screen.