Adjust `git close` to handle closing the current branch into dev/master
Run with `git close dev`, it will now take the current branch as
determined by running `git rev-parse --abbrev-ref HEAD` and try to see
if it's an issue branch. If it is, the script will check out `dev` as
requested, and then proceed as though you had called the script with an
issue number instead. Calling the script as `git close master` does the
same thing, but with `master` instead of `dev`.

This streamlines the workflow I use when finishing work on a new feature
or a bug fix, and removes the need for me to type the correct number,
which I don't always do.