Add newer terminal config options
The copy formatting option has grown since I originally set it, so
changing it to "none" here.

Don't want my terminal to detect URLs and try to make them clickable -
if I want to go to a URL from my terminal, I'll copy and paste it.

Trim block selection should remove trailing whitespace when copying text
from a large block of the terminal.

Tab width mode is for the terminal tabs, not for text; I think "equal"
is/was the existing behavior, and I'm used to it now, so keep it that
way.