Set maintenance.auto to false for my global git config
This was highlighted by @foone recently - essentially, git will
sometimes decide to perform maintenance tasks for a repository you're
working on, and will do them in the background. This is probably
intended to make life easier for the average developer, as it will make
sure their repositories get garbage-collected and such on a regular
basis, but I already run most of those things manually, and would prefer
to not have another background process fiddling with files without me
knowing, especially in my source controlled directories (even if it's
just stuff under .git).